cli: add a --secret flag to esc env set
This flag marks the value as secret by automatically wrapping it in a
call to `fn::secret`. These changes also autodetect secret-ish values
using the same method as the Pulumi CLI.

These changes also tweak the test harness so that the value of `run` is
a shell script rather than a single `esc` command. This allows for
richer tests by chaining commands.

Fixes #29.