encoding: fix an infinite loop
It was possible for the byte-offset calcuation in positionIndex.pos to
end up in an infinite loop if the column as calculated by `uniseg`
differed from the column as calculated by the YAML parser. These changes
eliminate that possibility by terminating the loop if there are no more
bytes in the line being processed.

These changes also add a small optimization to avoid dealing with
Unicode Text Segmentation for lines that are entirely ASCII.