esc, cli: helpers for well-known properties
These changes add helpers for fetching the environment variables and
temporary files defined by an environment. The former are sourced from
the `environmentVariables` top-level property and the latter from the
`files` top-level property. In both cases, nested properties with scalar
values are coerced to strings and nested properties with non-scalar
values are ignored.

Fixes #118.