eval: improve unknown/secret for complex literals
Do not treat object and array literals that include unknown or secret
elements as unknown or secret themselves. Instead, leave it up to the
consumer of the literal to handle unknownness and secretness
appropriately (i.e. by combining the unknownness and secretness of each
leaf value the consumer observes).

For example, the following object definition no is no longer unknown
during check or secret during eval:

```yaml
foo:
  bar: 42
  baz:
    fn::secret:
      ciphertext: AjIAEJFDIe==
```

Instead, it is partially-known during check:

```yaml
foo:
  bar: 42
  baz: [unknown]
```

And partially-secret during eval:

```yaml
foo:
  bar: 42
  baz: [secret]
```

But when passed to `fn::toJSON`, the result of the call would be unknown
in check or secret in eval b/c of the nested unknown/secret.

Fixes #164.
Fixes https://github.com/pulumi/pulumi-service/issues/16162.