eval: check for a nil import name
If an import declaration has a nil name, we can't process it any
further, so we can terminate early.

Also, change the "nil in interface" check in `exprPosition` to use
`IsZero` instead of `IsNil`, as the latter may panic if the value in the
interface is not a pointer (or other nillish type).