cli: prefer '@version' when env name is implied
`env diff` and `env version tag` both accept arguments that imply an
environment name. In the case of `env diff`, this is the version to use
for comparison; for `env version tag` it is the version to target with
the tag. In both cases, document that the syntax for these arguments is
`@version` (but continue to accept a bare version).