[cli] Stringify scalar secrets (#353)
`esc env set --secret` requires that its argument is a string literal.
While this is consistent with the requirements of the evaluator, it is
somewhat less than ergonomic for CLI users, who must surround numeric or
boolean values in both single and double quotes:

```console
esc env set my-env --secret foo.bar '"1234"'
```

These changes detect non-string scalar values passed to `esc env set --secret`
and automatically coerce these values to strings. Non-scalar values are
passed through to the evaluator as-is, which will report that the
argument must be a string literal.

Fixes #296.