eval: fix showSecrets in CheckEnvironment
These changes fix a bug in `CheckEnvironment` when `showSecrets` is
`false`. In this case, if the environment loader supplied a decrypter
for an imported environment, the imported environment's secrets would be
decrypted, violating the `showSecrets` contract. In order to address
this issue, these changes store and propagate the value of `showSecrets`
in the evaluator's context.