Remove dependency on `pulumi` `sdk/go/common/env` (#421)
This is the only use of `sdk/go/common/env` outside of `pulumi/pulumi`,
and it's easily removed by inlining the name of the relevant environment
variable. Since we already do this in a few other places (e.g. backend
URL lookup), this commit makes the change and breaks the dependency.