Add rotator providers (#432)
Introduces a new provider type and verb to support rotating static credentials.

Rotator providers have an additional state input which is used as a write-back target. The result of invoking Rotate is persisted back into the environment to this key. This provides a stable location to write to without accidentally clobbering interpolations that might be used for other inputs.

The new fn::rotate function type is used to invoke this new type of provider. It behaves similarly to existing providers during Open, but during rotation the evaluator will invoke the Rotate methods and collect their outputs, which are returned to the caller. The caller is expected to persist these updates back into the environment as a new revision.

```
# long form
fn::rotate:
  rotator: provider
  inputs:
    # any config the provider needs to use during rotation
  state:
    # persisted state

# short form
fn::rotate::provider:
  inputs:
    # any config the provider needs to use during rotation
  state:
    # persisted state
```

Closes pulumi/pulumi-service#24986