fix optional rotation state (#436)
rotation state was intended to be optional, so that it could be left off a new fn::rotate if not bootstrapping initial values, but this doesn't actually work currently. leaving state off does parse, but it ends up evaluating to an unknown value (since it's a missingExpr) which causes the evaluator to bail out.

this tries to fix that by treating a missing state key as null, and relaxing the rotator schema to always allow state to be null.