fix panic when importing an invalid environment (#465)
The import loader accidentally returns a nil environment if it fails to parse. Ordinarily this shouldn't happen because we don't let you save an environment if it fails validation. However, #449 introduced syntax errors for invalid fn calls, so some environments that previously incorrectly validated will now fail, and if you import one of these environments, we panic.