syntax: add a node for document trivia
These changes add a new node type, DocumentNode, to the syntax package.
This node allows concrete syntaxes to attach document-level syntactical
trivia (e.g. comments, metadata, etc.) that is not otherwise associated
with a node.

These changes also replace `*RotationResult` with `RotationResult`, as
the result is already nilable. This fixes an issue when running the
eval tests with `PULUMI_ACCEPT` set.

Fixes #260.