CI: reduce scope of permissions of test & lint jobs (#515)
The last release failed with a permission problem: https://github.com/pulumi/esc/actions/runs/14630282476

The publish-release workflow uses a subjob for lint. the lint job says it wants permissions: read-all but the parent workflow has a subset of permissions. I suspect that the models: read permission is new and didn't used to be included in read-all. Now that it is the parent permissions no longer cover what the child says it requires.

The lint/test jobs shouldn't need such broad permissions anyways, so this scopes them down to just being able to checkout the repo.