eval: tolerate syntax errors (#596)
We must tolerate syntax errors when loading an environment. Returning a
nil `EnvironmentDecl` prevents us from being able to analyze the
portions of the environment that were successfully parsed. This
interferes with analysis of environments that are only partially
erroneous. One important scenario this covers is autocompletion in the
face of syntax errors, which is common while editing an environment.